feat: add payment, contribution, guest user, validations and config

+++ src/validators/giftValidator.ts
@@ -4,9 +4,9 @@ export const createGiftSchema = z.object({
   name: z.string().min(3, 'O título deve ter pelo menos 3 caracteres'),
   description: z.string().min(3, 'A descrição deve ter pelo menos 3 caracteres'),
   photo: z.string().url('A URL da foto está inválida'),
-  totalValue: z.number().int().positive('O valor total deve ser positivo'),
+  totalValue: z.number().positive('O valor total deve ser positivo'),
   giftShares: z.number().int().positive('O número de cotas deve ser positivo'),
-  valuePerShare: z.number().int().positive('O valor por cota deve ser positivo'),
+  valuePerShare: z.number().positive('O valor por cota deve ser positivo'),
   categoryId: z.string().min(3, 'O ID da categoria está inválido'),
   giftListId: z.string().min(3, 'O ID da lista de presentes está inválido'),
 });
